categories get: move to hook to reuse from product
